move kv helpers to helpers_kv*.go
